pyproject.toml
[project]
name = "icml-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
package-mode = false
dependencies = [
    "fastcluster>=1.2.6",
    "igraph>=0.11.8",
    "ipdb>=0.13.13",
    "ipython>=5.2.0",
    "leidenalg>=0.10.2",
    "nbconvert>=7.16.6",
    "scanpy>=1.11.1",
    "scienceplots>=2.1.1",
    "scipy (==1.14.1)",
    "seaborn>=0.13.2",
    "ipykernel (>=6.29.5,<7.0.0)",
    "cell-gears (>=0.1.2,<0.2.0)",
    "torch-geometric (>=2.6.1,<3.0.0)",
    "numpy (==2.2.0)",
    "pandas (==2.2.2)",
    "tqdm (==4.66.5)",
    "scikit-learn (==1.5.1)",
    "networkx (==3.3)",
    "jupytext>=1.17.1",
]
